made user name not nullable

+++ src/api/routes.py
@@ -5,6 +5,7 @@ from flask import Flask, request, jsonify, url_for, Blueprint
 from api.models import db, User, Post, Favorites
 from api.utils import generate_sitemap, APIException
 from werkzeug.security import generate_password_hash
+from flask_jwt_extended import JWTManager, jwt_required
 
 api = Blueprint('api', __name__)
 app = Flask(__name__)
@@ -28,21 +29,19 @@ def signup_user():
     return jsonify({'message': 'registered successfully'}), 201  # HTTP status code for created
 
 
-@app.route('/login', methods=['POST'])
-def login_user(): 
-    auth = request.authorization
-    if not auth or not auth.username or not auth.password:  
-        return make_response('could not verify', 401, {'WWW.Authentication': 'Basic realm: "login required"'})    
-    
-    user = next((x for x in users if x.username == auth.username), None)
-    if not user:
-        return make_response('could not verify', 401, {'WWW.Authentication': 'Basic realm: "login required"'})
-    
-    if check_password_hash(user.password, auth.password):  
-        token = jwt.encode({'public_id': user.username, 'exp' : datetime.datetime.utcnow() + datetime.timedelta(minutes=30)}, app.config['SECRET_KEY'])
-        return jsonify({'token' : token.decode('UTF-8')})
+@api.route('/login', methods=['POST'])
+def login():
+    username = request.json.get("username", None)
+    password = request.json.get("password", None)
 
-    return make_response('could not verify',  401, {'WWW.Authentication': 'Basic realm: "login required"'})
+    # Perform authentication
+    user = User.query.filter_by(username=username).first()
+    if user is None or not user.check_password(password):
+        return jsonify({"msg": "Incorrect email or password"}), 401
+
+    # Generate access token
+    access_token = create_access_token(identity=username)
+    return jsonify(access_token=access_token)
 
 
 @api.route('/hello', methods=['POST', 'GET'])
@@ -56,26 +55,37 @@ def handle_hello():
 
 
 @api.route('/users/<int:id>', methods=['GET'])
+@jwt_required()
 def get_user_(id):
+# decorator on private routes
     user = User.query.filter_by(id=id).first()
     #serialized_users = [user.serialize() for user in user]
+    if not user :
+        return jsonify ({'message': 'user not found'})
     return jsonify(user.serialize())
 
 @api.route('/users/favorites', methods=['POST'])
+@jwt_required()
 def add_favorite():
     data = request.get_json()
     favorite = Favorites(
         user_id=data['user_id'],
         post_id=data['post_id'],
     )
-    
+    user = User.query.filter_by(id=id).first()
+    if not user :
+        return jsonify ({'message': 'user not found'})
     db.session.add(favorite)
     db.session.commit()
     return "ADD SUCCESS"
 
 @api.route('/users/favorites/<int:id>', methods=['DELETE'])
+@jwt_required()
 def delete_favorite(id):
     favorite = Favorites.query.get(id)
+    user = User.query.filter_by(id=id).first()
+    if not user :
+        return jsonify ({'message': 'user not found'})
     if favorite:
         db.session.delete(favorite)
         db.session.commit()
